Watchers - example
Example of how to use watch to detect changes being made to a data
variable and then take action each time a change is detected. In this
case I have an event on a button which increments the value of myNumber.
Each time this variable is updated "watch" detects this and sets the
value of myNumber back to 0 after a 3 second delay. This demonstrates
one of the potential pitfalls of using this method. A good example of
where this could be used is in a debounce or anywhere you would expect
very little thread blocking to occur in the watch function you are
writing.