Computed class
This shows a more complex method to manage dynamic class management.
Complex CSS logic can be handled with computed properties and then
passed into any element as an object which will display the CSS class in
the element based on the truthyness of a data variable. This can allow
for some very complex CSS state managment and updates from user input.