Emit using $event to update a variable stored in a prop
This is how you can communicate information in between the parent and child and back again. This is 2 way data binding in-between components.
A better way of handling state can be found when using "VueX". This depends on what sort of application is being built and props are certainly a great tool to use in a huge number of applications